infoInq 공통 검색으로 수정
