+++ front/src/pages/infoInq/EmpListJson.json
@@ -1,21 +1,106 @@
 {
-	"menuName": "EmpList",
-	"queryId": "infoInqMapper.retrieveEmpList",
+    "menuName": "EmpList",
+    "queryId": "infoInqMapper.retrieveEmpList",
     "keyColumn": "empno",
-    "tableColumns": 
-	[
-        { "key": "empno", "value": "사번", "width": 120, "alignment" : "center" },
-        { "key": "empFlnm", "value": "성명", "width": 100 , "alignment" : "center"},
-        { "key": "jbpsNm", "value": "직위", "width": 200, "alignment" : "center" },
-        { "key": "deptNm", "value": "소속", "width": 170, "alignment" : "center" },
-        { "key": "telno", "value": "전화번호", "width": 170, "alignment" : "center"  },
-        { "key": "eml", "value": "이메일", "width": 200, "alignment" : "center"  },
-        { "key": "hodfSttsNm", "value": "재직상태", "width": 100,  "alignment" : "center", "border" : "1px" },
-        { "key": "bankNm", "value": "은행코드", "width": 100,  "alignment" : "center", "border" : "1px" , "visible" : true},
-        { "key": "actNo", "value": "계좌번호", "width": 100,  "alignment" : "center", "border" : "1px", "visible" : true }
-	],
-    "searchParams":
-	{
+    "tableColumns": [
+        {
+            "key": "empno",
+            "value": "사번",
+            "width": 120,
+            "alignment": "center"
+        },
+        {
+            "key": "empFlnm",
+            "value": "성명",
+            "width": 100,
+            "alignment": "center"
+        },
+        {
+            "key": "jbpsNm",
+            "value": "직위",
+            "width": 200,
+            "alignment": "center"
+        },
+        {
+            "key": "deptNm",
+            "value": "소속",
+            "width": 170,
+            "alignment": "center"
+        },
+        {
+            "key": "telno",
+            "value": "전화번호",
+            "width": 170,
+            "alignment": "center"
+        },
+        {
+            "key": "eml",
+            "value": "이메일",
+            "width": 200,
+            "alignment": "center"
+        },
+        {
+            "key": "hodfSttsNm",
+            "value": "재직상태",
+            "width": 100,
+            "alignment": "center",
+            "border": "1px"
+        },
+        {
+            "key": "bankNm",
+            "value": "은행코드",
+            "width": 100,
+            "alignment": "center",
+            "border": "1px",
+            "visible": true
+        },
+        {
+            "key": "actNo",
+            "value": "계좌번호",
+            "width": 100,
+            "alignment": "center",
+            "border": "1px",
+            "visible": true
+        }
+    ],
+    "searchInfo": {
+        "searchParams": {
+            "insertButton": false
+        },
+        "commonCd": "VTW001",
+        "textBoxItem": [
+            {
+                "className": "empNoItem",
+                "placeholder": "사번",
+                "name": "empno"
+            },
+            {
+                "className": "empNoItem",
+                "placeholder": "성명",
+                "name": "empFlnm"
+            },
+            {
+                "className": "empTel",
+                "placeholder": "전화번호",
+                "name": "telNo"
+            }
+        ],
+        "selectBoxItem": [
+            {
+                "commonCd": "VTW003",
+                "placeholderText": "재직여부",
+                "name": "hodfSttsNm"
+            },
+            {
+                "commonCd": "VTW001",
+                "placeholderText": "직위",
+                "name": "jbpsNm"
+            },
+            {
+                "commonCd": "VTW001",
+                "placeholderText": "소속",
+                "name": "deptNm"
+            }
+        ]
     }
-    
 }
